+++ README.md
@@ -61,6 +61,7 @@ Stempo는 뇌성마비 및 후천적 장애 환자들의 리듬기반 재활 훈
     - **config**: API 요청 및 응답에 대한 설정 클래스들이 포함됩니다.
     - **controller**: HTTP 요청을 처리하고, 서비스 계층에 비즈니스 로직을 전달하며, 처리된 결과를 응답으로 반환하는 컨트롤러 클래스들이 포함됩니다.
     - **exception**: API 요청 처리 중 발생할 수 있는 예외를 처리하고 클라이언트에 적절한 응답을 반환하는 클래스들이 포함됩니다.
+    - **interceptor**: HTTP 요청을 가로채어 처리하는 인터셉터 클래스들이 포함됩니다.
     - **support**: YAML 파일을 로딩하거나 파싱하는 등의 지원 기능을 제공하는 유틸리티 클래스들이 포함됩니다.
 
 - `stempo-application`: 비즈니스 로직을 처리하고 도메인 객체와 상호작용하는 모듈입니다.
@@ -75,7 +76,6 @@ Stempo는 뇌성마비 및 후천적 장애 환자들의 리듬기반 재활 훈
     - **config**: 인증 및 보안 설정을 관리하는 클래스들이 포함됩니다.
     - **exception**: 인증 및 권한 부여 과정에서 발생하는 예외를 처리하는 클래스들이 포함됩니다.
     - **filter**: HTTP 요청을 가로채어 인증 및 권한 부여를 처리하는 필터 클래스들이 포함됩니다.
-    - **handler**: 인증과 관련된 로깅을 처리하는 인터셉터 클래스들이 포함됩니다.
     - **model**: 사용자 인증 정보 및 관련 모델 객체들이 포함됩니다.
     - **util**: 인증 및 권한 관련 유틸리티 클래스들이 포함됩니다.
