+++ stempo-api/src/main/java/com/stempo/interceptor/ApiLoggingInterceptor.java
@@ -1,4 +1,4 @@
-package com.stempo.handler;
+package com.stempo.interceptor;
 
 import com.stempo.util.HttpReqResUtils;
 import jakarta.servlet.http.HttpServletRequest;
